heavy changes and refactoring
i now use a matrix of bool values to keep track of which cells that have been visited in the DFS.
A DFS search is conducted for each node to find the connections to other nodes and their distance.

further more prims algorithm has been implemented to find the minimal spanning tree.

as of now, the distance produced is wrong. the value is to high.