Complete rework of code
The solution is simplified, we now only create a heatMap and traverse the matrix directly instead of fist creating an adjacency matrix.
The key difference with this solution is that we keep track of not only if a node has been visited, we also check if it has been visited with a specific state, that is which position (x and y coordinates), the x and y movement direction taken and how many steps in that direction was taken.  This is done with a 5d-bool array.